requirements.txt
torch>=1.12.0
transformers>=4.30.0
arabert-preprocessor>=1.0.0
scikit-learn>=1.2.0
pandas>=1.4.0
joblib>=1.0.0
